add PE binary cataloger
