+++ syft/pkg/cataloger/internal/pe/test-fixtures/image-net8-app
@@ -1,0 +1,1 @@
+../../../dotnet/test-fixtures/image-net8-app
