+++ syft/pkg/cataloger/internal/pe/test-fixtures/image-net8-app-single-file
@@ -1,0 +1,1 @@
+../../../dotnet/test-fixtures/image-net8-app-single-file
